plus, minus, degree for polynomial
